Fix validations, tidy up css and move duplicate function

+++ src/routes/team/deleteModal.svelte
@@ -3,23 +3,20 @@
   import { PUBLIC_URI } from "$env/static/public";
   import Modal from "$lib/modal/globalModal.svelte";
   import { deleteMode, requestStatus, tableRefresh } from "$lib/stores/stores";
+  import { getUserFromSessionStorage } from "$lib/customFunctions";
+
+  const loggedInUser: any = getUserFromSessionStorage(); //get the logged in user from sessionStorage
 
   export let showModal = false;
   export let teamName: any;
   let msg = "";
-  const loggedInUser: any = {};
-
-  if (typeof sessionStorage !== "undefined") {
-    const userLoggedIn = sessionStorage.getItem("userLoggedIn");
-    if (userLoggedIn !== null) {
-      for (const [key, value] of Object.entries(JSON.parse(userLoggedIn))) {
-        loggedInUser[key] = value;
-      }
-    }
-  }
 
   async function deleteTeam() {
     validatePermissions();
+    if (teamName === "Super") {
+      msg = "Cannot delete this team!";
+      return;
+    }
     await fetch(`${PUBLIC_URI}/teams?team_name=${teamName}`, {
       method: "DELETE",
     })
@@ -33,7 +30,7 @@
         showModal = false;
         requestStatus.set("success");
         tableRefresh.set(true); //fresh page on success
-        })
+      })
       .catch((err) => {
         console.error(err);
         throw err; // Re-throw the error to propagate it to the caller
@@ -59,9 +56,8 @@
     <p class="errorMessage">{msg}</p>
   {/if}
   <div class="flex flex-col mt-2">
-    <button
-      class="text-white bg-red-600 hover:bg-red-800 focus:outline-none font-medium rounded-lg text-sm px-5 py-2.5 text-center mr-2 mb-2"
-      on:click={() => deleteTeam()}>Delete Team</button
+    <button class="deleteModalButton" on:click={() => deleteTeam()}
+      >Delete Team</button
     >
   </div>
 </Modal>
